feat: migrate to using tsup for building
