--- scripts/testBuild.ts
@@ -1,9 +1,0 @@
-import * as esUtils from '../build/esm'
-
-const cjsUtils = require('../build/cjs')
-
-console.log(`------- cjsUtils -------`)
-console.log(cjsUtils)
-
-console.log(`------- esUtils -------`)
-console.log(esUtils)
